% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{default_init_example}
\alias{default_init_example}
\title{Example initial parameters for the full multimix model}
\format{
An object of class \code{numeric} of length 12.
}
\usage{
default_init_example
}
\description{
Example initial parameters for the full multimix model
}
\keyword{datasets}
