% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{default_init_example_lite}
\alias{default_init_example_lite}
\title{Example initial parameters for the lite multimix model}
\format{
An object of class \code{numeric} of length 11.
}
\usage{
default_init_example_lite
}
\description{
Example initial parameters for the lite multimix model
}
\keyword{datasets}
