% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{get_early_phase}
\alias{get_early_phase}
\title{Get early phase temporal component (density)}
\usage{
get_early_phase(time, thalf, eta = 1, gamma = 0)
}
\arguments{
\item{time}{Numeric vector of time values.}

\item{thalf}{Half-life parameter.}

\item{eta}{Exponent of time (nu in decompos notation). Default 1.}

\item{gamma}{Shape parameter (m in decompos notation). Default 0.}
}
\value{
Numeric vector of early phase density values.
}
\description{
Get early phase temporal component (density)
}
