% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_functions.R
\name{get_late_phase}
\alias{get_late_phase}
\title{Get late phase temporal component (hazard)}
\usage{
get_late_phase(time, thalf, eta, gamma = 0)
}
\arguments{
\item{time}{Numeric vector of time values.}

\item{thalf}{Half-life parameter.}

\item{eta}{Exponent of time (nu in decompos notation).}

\item{gamma}{Shape parameter (m in decompos notation). Default 0.}
}
\value{
Numeric vector of late phase hazard values.
}
\description{
Get late phase temporal component (hazard)
}
