% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{lower_bounds_example}
\alias{lower_bounds_example}
\title{Example lower bounds for the full multimix model}
\format{
An object of class \code{numeric} of length 12.
}
\usage{
lower_bounds_example
}
\description{
Example lower bounds for the full multimix model
}
\keyword{datasets}
