% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{lower_bounds_example_lite}
\alias{lower_bounds_example_lite}
\title{Example lower bounds for the lite multimix model}
\format{
An object of class \code{numeric} of length 11.
}
\usage{
lower_bounds_example_lite
}
\description{
Example lower bounds for the lite multimix model
}
\keyword{datasets}
