% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{upper_bounds_example}
\alias{upper_bounds_example}
\title{Example upper bounds for the full multimix model}
\format{
An object of class \code{numeric} of length 12.
}
\usage{
upper_bounds_example
}
\description{
Example upper bounds for the full multimix model
}
\keyword{datasets}
