% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{upper_bounds_example_lite}
\alias{upper_bounds_example_lite}
\title{Example upper bounds for the lite multimix model}
\format{
An object of class \code{numeric} of length 11.
}
\usage{
upper_bounds_example_lite
}
\description{
Example upper bounds for the lite multimix model
}
\keyword{datasets}
